crypto: use u32 for flag in quic_crypto_set_cipher()
The flag argument in quic_crypto_set_cipher() and quic_crypto_set_secret()
is used as a bitmask to indicate different installation directions and key
types. Using u8 unnecessarily restricts the available flag space and is
inconsistent with the rest of the QUIC crypto code, where flags are
typically represented as u32.

Update the type of flag from u8 to u32 in both functions and their
declarations to allow future expansion and maintain consistency with other
flag definitions.

+++ modules/net/quic/crypto.c
@@ -637,7 +637,7 @@ out:
 }
 EXPORT_SYMBOL_GPL(quic_crypto_decrypt);
 
-int quic_crypto_set_cipher(struct quic_crypto *crypto, u32 type, u8 flag)
+int quic_crypto_set_cipher(struct quic_crypto *crypto, u32 type, u32 flag)
 {
 	struct quic_cipher *cipher;
 	void *tfm;
@@ -711,7 +711,7 @@ err:
 }
 
 int quic_crypto_set_secret(struct quic_crypto *crypto, struct quic_crypto_secret *srt,
-			   u32 version, u8 flag)
+			   u32 version, u32 flag)
 {
 	struct quic_cipher *cipher;
 	int err;
