+++ modules/net/quic/crypto.h
@@ -57,9 +57,9 @@ struct quic_crypto {
 };
 
 int quic_crypto_set_secret(struct quic_crypto *crypto, struct quic_crypto_secret *srt,
-			   u32 version, u8 flag);
+			   u32 version, u32 flag);
 int quic_crypto_get_secret(struct quic_crypto *crypto, struct quic_crypto_secret *srt);
-int quic_crypto_set_cipher(struct quic_crypto *crypto, u32 type, u8 flag);
+int quic_crypto_set_cipher(struct quic_crypto *crypto, u32 type, u32 flag);
 int quic_crypto_key_update(struct quic_crypto *crypto);
 
 int quic_crypto_encrypt(struct quic_crypto *crypto, struct sk_buff *skb);
